+++ src/clojure_lsp/feature/code_lens.clj
@@ -20,18 +20,30 @@
 (defn ^:private test-references->string [references]
   (references->string references " test"))
 
-(defn reference-code-lens [uri]
-  (let [analysis (get @db/db :analysis)
-        excluded-vars (set/union #{'clojure.test/deftest}
-                                 (get-in @db/db [:settings :linters :unused-public-var :exclude-when-defined-by] #{}))]
-    (->> (q/find-vars analysis (shared/uri->filename uri) true)
+(defn ^:private var-definitions-lens [filename settings analysis]
+  (let [excluded-vars (set/union #{'clojure.test/deftest}
+                                 (get-in settings [:linters :unused-public-var :exclude-when-defined-by] #{}))]
+    (->> (q/find-var-definitions analysis filename true)
          (remove #(or (contains? excluded-vars (:defined-by %))
                       (and (= 'clojure.core/defrecord (:defined-by %))
                            (or (string/starts-with? (str (:name %)) "->")
-                               (string/starts-with? (str (:name %)) "map->")))))
-         (map (fn [var]
-                {:range (shared/->range var)
-                 :data  [uri (:name-row var) (:name-col var)]})))))
+                               (string/starts-with? (str (:name %)) "map->"))))))))
+
+(defn ^:private keyword-definitions-lens
+  [filename settings analysis]
+  (let [excluded-vars (get-in settings [:linters :unused-public-var :exclude-when-defined-by] #{})]
+    (->> (q/find-keyword-definitions analysis filename)
+         (remove #(contains? excluded-vars (:reg %))))))
+
+(defn reference-code-lens [uri]
+  (let [analysis (:analysis @db/db)
+        settings (:settings @db/db)
+        filename (shared/uri->filename uri)]
+    (->> (concat (var-definitions-lens filename settings analysis)
+                 (keyword-definitions-lens filename settings analysis))
+         (map (fn [element]
+                {:range (shared/->range element)
+                 :data  [uri (:name-row element) (:name-col element)]})))))
 
 (defn test-reference? [source-path {:keys [filename]}]
   (and source-path